Move PSBT serialize code to submodule
Now that we have separated the serialization logic out we can move it to
a `serialize` module.
