+++ bitcoin/src/psbt/serialize/error.rs
@@ -4,10 +4,10 @@ use core::fmt;
 
 use internals::write_err;
 
+use super::raw;
 use crate::bip32::Xpub;
 use crate::consensus::encode;
 use crate::prelude::Box;
-use crate::psbt::raw;
 use crate::transaction::Transaction;
 
 /// Enum for marking psbt hash error.
