+++ bitcoin/src/psbt/serialize/map/mod.rs
@@ -5,7 +5,7 @@ mod input;
 mod output;
 
 use crate::prelude::Vec;
-use crate::psbt::raw;
+use crate::psbt::serialize::raw;
 use crate::psbt::serialize::Serialize;
 
 #[rustfmt::skip]                // Keep public re-exports separate.
